revert changes to upload
