+++ packages/hadron-build/commands/write-assets.js
@@ -3,25 +3,24 @@
 const path = require('path');
 const fs = require('fs');
 const Target = require('../lib/target');
-const { generateVersionsForAssets } = require('../lib/assets');
 
 const cli = require('mongodb-js-cli')('hadron-build:download');
 
 const command = 'write-assets [options]';
 
-const describe = 'Write the build assets to console or a file';
+const describe = 'Write development build assets to console or a file';
 
 const builder = {
   dir: {
     description: 'Project root directory',
     default: process.cwd(),
   },
-  url: {
+  'download-url': {
     description: 'Base download url of the asset',
   },
-  version: {
-    description: 'Target version',
-    default: undefined,
+  'build-id': {
+    description: 'Build ID',
+    default: (process.env.DEV_VERSION_IDENTIFIER ?? '').replaceAll('_', ''),
   },
   out: {
     description: 'Output file path',
@@ -31,22 +30,31 @@ const builder = {
 
 const handler = function handler(argv) {
   cli.argv = argv;
-  const version = (
-    argv.version ? argv.version : new Target(argv.dir).version
-  ).replace(/^v/, '');
+  const version = new Target(argv.dir).version;
 
-  const assets = generateVersionsForAssets(
-    Target.getAssetsForVersion(argv.dir, version),
-    version,
-    argv.url
-  );
+  const assets = Target.getAssetsForVersion(argv.dir, version);
+  const mappedAssets = {
+    id: argv.buildId,
+    tag_name: `v${version}`,
+    name: version,
+    draft: true,
+    created_at: new Date(),
+    assets: assets.flatMap(({ assets: items }) => items.map(x => ({
+      url: `${argv.downloadUrl}/${x.name}`,
+      name: x.name,
+      id: x.name,
+      label: x.name,
+    }))),
+    body: 'This is a development release and all the changes can be found on https://github.com/mongodb-js/compass'
+  };
+
   if (argv.out) {
     const out = path.resolve(process.cwd(), argv.out);
     fs.mkdirSync(path.dirname(out), { recursive: true });
-    fs.writeFileSync(out, JSON.stringify(assets, null, 2));
+    fs.writeFileSync(out, JSON.stringify(mappedAssets, null, 2));
   } else {
     // eslint-disable-next-line no-console
-    console.log(JSON.stringify(assets, null, 2));
+    console.log(JSON.stringify(mappedAssets, null, 2));
   }
 };
